cc-2188: Download action for all media file.
Download action were added in two placese.

1. Playlist Builder. jjmenu now showing 'Download'.
2. Calendar-> Show Content. Click on the file, you will see
Downlaod menu.

Limitation:

File name is set to original file name.
On the ticket it said, file format should be like
"show_name-12012011-2130.mp3". However, in my opinion,
this is only possible, if a user downloads it from
calender->show content. We know what the show names and etc,
but, if the user downloads it from Playlist Builder, we cannot
retrive such information. To be consistent in both area, I just
set the file name to original file name. This can be easily
modified in the future.