CC-2436: Save pulldown settings
Fixing the issue of "show entries" dropdown setting not being saved.
After some investigation, it seems like we're destroying the calendar
everytime we modify the dropdown value, and create a new one,
which explains why the new setting is not being saved.

Fixed by moving the code that handles the updating to the place after
we create the new calendar. This makes the code cleaner as well.